feat: integrate tests of vt for both webgl and webgpu (#2722)
* fix: fixing spec of default-icon in webgpu

* chore: change vt specs to gpu files

* fix: missed wgsl files

* fix: fix missing bus on linux

* fix: trying to fix dbus

* fix: fix transform-to-gpu.js

* fix transform-to-gpu.js

* remove dbus in config.yml

* fixing vt specs

+++ packages/vt/src/layer/plugins/painters/wgsl/marker_vert.wgsl
@@ -321,7 +321,7 @@ fn main(vertexInput: VertexInput) -> VertexOutput {
     output.vHalo = f32(vertexInput.aShape.w) % 2.0;
     output.vTextSize = myTextSize;
 #ifdef ENABLE_COLLISION
-    output.vOpacity = vertexInput.aOpacity / 255.0;
+    output.vOpacity = f32(vertexInput.aOpacity) / 255.0;
 #else
     output.vOpacity = 1.0;
 #endif
